Gusturoi-1-Little Refactorization
