+++ src/main/java/com/sef/App.java
@@ -1,10 +1,36 @@
 package com.sef;
 
-import com.sef.gui.GUI;
+import java.io.IOException;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
 
-public class App {
+public class App extends Application {
+
+  private static Scene scene;
+
+  @Override
+  public void start(Stage stage) throws IOException {
+    scene = new Scene(loadFXML("login"), 1600, 900);
+    stage.setTitle("International Cuisine");
+    stage.setScene(scene);
+    stage.show();
+  }
+
+  public static void setRoot(String fxml) throws IOException {
+    scene.setRoot(loadFXML(fxml));
+  }
+
+  private static Parent loadFXML(String fxml) throws IOException {
+    FXMLLoader fxmlLoader = new FXMLLoader(
+      App.class.getResource("/com/sef/" + fxml + ".fxml")
+    );
+    return fxmlLoader.load();
+  }
 
   public static void main(String[] args) {
-    GUI.main(args);
+    launch();
   }
 }
