+++ src/main/java/com/sef/controllers/LoginController.java
@@ -1,6 +1,6 @@
-package com.sef.gui.controllers;
+package com.sef.controllers;
 
-import com.sef.gui.GUI;
+import com.sef.App;
 import java.io.IOException;
 import javafx.fxml.FXML;
 import javafx.scene.control.PasswordField;
@@ -17,6 +17,8 @@ public class LoginController {
 
   @FXML
   public TextField usernameField;
+
+  private String userName;
 
   @FXML
   public void handleLoginButtonAction() {
@@ -37,11 +39,13 @@ public class LoginController {
 
     if (username.equals("member") && password.equals("member")) {
       loginMessage.setText("Logged in as member!");
+      userName = username;
       return;
     }
 
     if (username.equals("admin") && password.equals("admin")) {
       loginMessage.setText("Logged in as admin!");
+      userName = username;
       return;
     }
 
@@ -50,6 +54,10 @@ public class LoginController {
 
   @FXML
   private void switchToRegister() throws IOException {
-    GUI.setRoot("register");
+    App.setRoot("register");
+  }
+
+  public String getUserName() {
+    return userName;
   }
 }
