--- src/main/java/com/sef/gui/GUI.java
@@ -1,36 +1,0 @@
-package com.sef.gui;
-
-import java.io.IOException;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class GUI extends Application {
-
-  private static Scene scene;
-
-  @Override
-  public void start(Stage stage) throws IOException {
-    scene = new Scene(loadFXML("login"), 1600, 900);
-    stage.setTitle("International Cuisine");
-    stage.setScene(scene);
-    stage.show();
-  }
-
-  public static void setRoot(String fxml) throws IOException {
-    scene.setRoot(loadFXML(fxml));
-  }
-
-  private static Parent loadFXML(String fxml) throws IOException {
-    FXMLLoader fxmlLoader = new FXMLLoader(
-      GUI.class.getResource("/com/sef/" + fxml + ".fxml")
-    );
-    return fxmlLoader.load();
-  }
-
-  public static void main(String[] args) {
-    launch();
-  }
-}
