--- src/main/java/com/sef/gui/controllers/PrimaryController.java
@@ -1,13 +1,0 @@
-package com.sef.gui.controllers;
-
-import com.sef.gui.GUI;
-import java.io.IOException;
-import javafx.fxml.FXML;
-
-public class PrimaryController {
-
-  @FXML
-  private void switchToSecondary() throws IOException {
-    GUI.setRoot("secondary");
-  }
-}
