+++ src/main/resources/com/sef/register.fxml
@@ -11,7 +11,7 @@
   prefWidth="600"
   prefHeight="400"
   xmlns="http://javafx.com/javafx/8.0.171"
-  fx:controller="com.sef.gui.controllers.RegisterController"
+  fx:controller="com.sef.controllers.RegisterController"
 >
    <padding>
        <Insets bottom="25.0" left="25.0" right="25.0" top="25.0" />
