--- src/main/resources/com/sef/secondary.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.Button?>
-<?import javafx.geometry.Insets?>
-<VBox
-  alignment="CENTER"
-  spacing="20.0"
-  xmlns="http://javafx.com/javafx/8.0.171"
-  xmlns:fx="http://javafx.com/fxml/1"
-  fx:controller="com.sef.gui.controllers.SecondaryController"
->
-    <children>
-        <Label text="Secondary View" />
-        <Button
-      fx:id="secondaryButton"
-      text="Switch to Primary View"
-      onAction="#switchToPrimary"
-    />
-    </children>
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-    </padding>
-</VBox>
